:recycle: Factor out signed url tests
